requirements.txt
# pip install -r requirements.txt

tqdm
parso>=0.2.0
pyyaml
argparse
pylabel
pyfiglet
termcolor